Refactor: Remove map-constant values, apply code analysis recommendations, move display-only values to a separate class.
This is an initial refactor to move values that are constant for the map
(catcher width, clock rate, frame time, playfield width) out of
individual objects and at least at the moment, into method parameters.

This isn't perfect and almost makes things messier. I believe that using
a data class for the values might be a good and generalizable approach.

Along with the above, linting errors/warnings have been fixed and all
display-only values have been moved to CatchDisplayData.